Added a pasword schema, altered main
Added a password generating schema to create passwords of a specific word count that also included numbers and some capital letters. This was pretty much to be able to quickly generate passwords that met the requirements for many online services. If you're devious you can run this script as a shell command ...